Configurable number of simultaneous downloads.
