+++ src/keyMap.js
@@ -1,0 +1,8 @@
+export default {
+    'control+p': 'togglePause',
+    'control+t': 'putTimestamp',
+    'control+s': 'saveNote',
+    'control+shift+right': 'videoForward',
+    'control+shift+left': 'videoBackward',
+    'control+shift+s': 'save', // Saves to both local storage and dropbox. Just mentioning.
+};
